+++ src/receiver.h
@@ -30,4 +30,4 @@ channel_t knob_s2;
 
 void receiver_init(uart_inst_t* uart_id, int tx_pin, int rx_pin);
 
-int receiver_check_if_disconnected();
+uint8_t receiver_check_if_disconnected();
